+++ CHANGELOG.md
@@ -4,15 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [21.1.6]
+* Resolved issue with structure sets not being loaded properly
+
+## [21.1.5]
+
+### Added
+* Support structure sets in the `pregen` dimension construction type
+
 ## [21.1.4]
 
-- Added Chinese translation
+### Added
+* Added Chinese translation
 
 ## [21.1.3]
 
 ### Added
 * Added ru_ru translation (thanks @BazZziliuS)
-* 
+
 ### Fixed
 * Fixed players returning from the Nether getting sent to base spawn instead of the portal they went through